MNIST_CNN/requirements.txt
absl-py==2.1.0
astor==0.8.1
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.38.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.1.0
nltk==3.4.4
numpy==1.16.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
protobuf==3.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.2.1
six==1.16.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==2.3.0
tk==0.1.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
